Add skeleton for Question 2 ML modeling
The skeleton serves as a first-look approach to the models that
seemed to perform the best from reviewed literature - the attempt
was to reimplement in Python via `sklearn`. There are lots of
refinement and fine tuning to be done (from clarification of
requirements to correct cross validation approach for this problem),
but the various classifiers are implemented with defaults. This could
serve as a high-level approach with basic, preliminary results that
appear to be on the right track and within expectations.